backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
jinja2
bcrypt
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
aiofiles
requests
websockets